JavaScript (v3): CloudWatch Logs Wkflw - Refactor dateRangeGenerator
